Added first tests to Sepsis Check
Extract relevant patient data from json: for sofa calculation and sepsis check + testing of some of the functions

The first timestep of the json has feature-names in all CAPS. I manually removed it to continue with the tests. (The original can still be found in ./data/patient)
The values of features is the same for each timestep? i.e. everytime Bilirubin is measured, its value is 0.4.

What I didnt find so far/is not in the data:
    - Cardiovascular: 'mean_arterial_pressure: float' & 'catecholamine: Optional[Catecholamine]' -> 'compute_score_for_cardiovascular_system'
    - Respiration - PaO2/FiO2 (mmHg): 'partial_pressure_of_oxygen: float' & 'is_mechanically_ventilated: bool' -> 'compute_score_for_respiratory_system'
    - GCS -> we will not use it
    - Urine output (mL/d): 'urine_output: float' -> 'compute_score_for_kidneys' -> is optional but probably useful and important
    - Feature for administered Antibiotics