Created .py for Sepsis Check and a notebook testing its functions.
SepsisCheck should now work as intended.

I also started with 'json_exploration.ipynb'
Right now trying to extract as much Features from the patient.jsons as possible through a combination of json structured data and extracting from text using spacy-models (med7, which i instsalled in a seperate environment on my pc) and regex. But i do not upload because of concerns with the patient data in prints?

sepsis_check.ipynb is now obsolete

Maybe move the files in folders later on, but may need slight modification because of paths.